Give explicit type to all literals
